Move the vscode extension configuration to its own extension.ts file to not conflict with the webExtension configuration.
